refactor: 删除StateProperty类并将其职能移交给StateInfo
